+++ ios/MullvadVPN/Coordinators/App/ApplicationCoordinator.swift
@@ -104,6 +104,7 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
         router = ApplicationRouter(self)
 
         addTunnelObserver()
+        setShowAccountButton(for: tunnelManager.deviceState)
     }
 
     func start() {
@@ -123,6 +124,9 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
         completion: @escaping (Coordinator) -> Void
     ) {
         switch route {
+        case .account:
+            presentAccount(animated: animated, completion: completion)
+
         case let .settings(subRoute):
             presentSettings(route: subRoute, animated: animated, completion: completion)
 
@@ -162,7 +166,7 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
                 endHorizontalFlow(animated: context.isAnimated, completion: completion)
                 context.dismissedRoutes.forEach { $0.coordinator.removeFromParent() }
 
-            case .selectLocation, .settings:
+            case .selectLocation, .account, .settings:
                 let coordinator = dismissedRoute.coordinator as! Presentable
 
                 coordinator.dismiss(animated: context.isAnimated, completion: completion)
@@ -292,9 +296,9 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
         }
     }
 
-    private func didDismissSettings(_ reason: SettingsDismissReason) {
+    private func didDismissAccount(_ reason: AccountDismissReason) {
         if isPad {
-            router.dismissAll(.settings, animated: true)
+            router.dismiss(.account, animated: true)
 
             if reason == .userLoggedOut {
                 router.dismissAll(.primary, animated: true)
@@ -306,7 +310,7 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
                 continueFlow(animated: false)
             }
 
-            router.dismissAll(.settings, animated: true)
+            router.dismiss(.account, animated: true)
         }
     }
 
@@ -542,6 +546,35 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
         return selectLocationCoordinator
     }
 
+    private func presentAccount(animated: Bool, completion: @escaping (Coordinator) -> Void) {
+        let accountInteractor = AccountInteractor(
+            storePaymentManager: storePaymentManager,
+            tunnelManager: tunnelManager
+        )
+
+        let coordinator = AccountCoordinator(
+            navigationController: CustomNavigationController(),
+            interactor: accountInteractor
+        )
+
+        coordinator.didFinish = { [weak self] coordinator, reason in
+            self?.didDismissAccount(reason)
+        }
+
+        coordinator.start(animated: animated)
+
+        presentChild(
+            coordinator,
+            animated: animated,
+            configuration: ModalPresentationConfiguration(
+                preferredContentSize: preferredFormSheetContentSize,
+                modalPresentationStyle: .formSheet
+            )
+        ) {
+            completion(coordinator)
+        }
+    }
+
     private func presentSettings(
         route: SettingsNavigationRoute?,
         animated: Bool,
@@ -559,8 +592,8 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
             interactorFactory: interactorFactory
         )
 
-        coordinator.didFinish = { [weak self] coordinator, reason in
-            self?.didDismissSettings(reason)
+        coordinator.didFinish = { [weak self] coordinator in
+            self?.router.dismissAll(.settings, animated: true)
         }
 
         coordinator.willNavigate = { [weak self] coordinator, from, to in
@@ -598,6 +631,7 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
 
     private func deviceStateDidChange(_ deviceState: DeviceState) {
         splitViewController.preferredDisplayMode = deviceState.splitViewMode
+        setShowAccountButton(for: deviceState)
 
         switch deviceState {
         case let .loggedIn(accountData, _):
@@ -606,7 +640,6 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
             if !accountData.isExpired {
                 router.dismiss(.outOfTime, animated: true)
             }
-
         case .revoked:
             cancelOutOfTimeTimer()
             router.present(.revoked, animated: true)
@@ -614,6 +647,16 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
         case .loggedOut:
             cancelOutOfTimeTimer()
         }
+    }
+
+    private func setShowAccountButton(for deviceState: DeviceState) {
+        var show = false
+        if case .loggedIn = deviceState {
+            show = true
+        }
+
+        primaryNavigationContainer.setShowAccountButton(show)
+        secondaryNavigationContainer.setShowAccountButton(show)
     }
 
     // MARK: - Out of time
@@ -651,8 +694,8 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
 
     // MARK: - Deep link
 
-    func showAccountSettings() {
-        router.present(.settings(.account))
+    func showAccount() {
+        router.present(.account)
     }
 
     // MARK: - UISplitViewControllerDelegate
@@ -689,6 +732,13 @@ final class ApplicationCoordinator: Coordinator, Presenting, RootContainerViewCo
     }
 
     // MARK: - RootContainerViewControllerDelegate
+
+    func rootContainerViewControllerShouldShowAccount(
+        _ controller: RootContainerViewController,
+        animated: Bool
+    ) {
+        router.present(.account, animated: animated)
+    }
 
     func rootContainerViewControllerShouldShowSettings(
         _ controller: RootContainerViewController,