+++ ios/CHANGELOG.md
@@ -25,6 +25,7 @@ Line wrap the file at 100 chars.                                              Th
 ## [Unreleased]
 ### Added
 - Add search functionality to location selection view.
+- Add a dedicated account button on the main view and remove it from settings.
 
 ### Changed
 - Changed key rotation interval from 4 to 14 days.